+++ files/srcs/errors.c
@@ -6,7 +6,7 @@
 /*   By: mmateo-t <mmateo-t@student.42madrid>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/12/06 11:45:39 by mmateo-t          #+#    #+#             */
-/*   Updated: 2022/01/03 12:59:09 by mmateo-t         ###   ########.fr       */
+/*   Updated: 2022/01/03 17:56:54 by mmateo-t         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -20,5 +20,8 @@ void	throw_error(const char *error)
 void	check_args(int argc)
 {
 	if (argc != 1)
+	{
 		throw_error("Numbers of arguments invalid");
+		exit(EXIT_FAILURE);
+	}
 }
