+++ files/srcs/signals.c
@@ -6,19 +6,18 @@
 /*   By: mmateo-t <mmateo-t@student.42madrid>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/12/17 12:22:42 by mmateo-t          #+#    #+#             */
-/*   Updated: 2021/12/17 13:19:42 by mmateo-t         ###   ########.fr       */
+/*   Updated: 2022/01/03 13:39:41 by mmateo-t         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../includes/minishell.h"
 
+//FIXME: ^C still appears in command line
+
 void sigint_handler(int sig)
 {
-	//printf("SIGNAL: %i", sig);
-	rl_replace_line("", 1);
-	rl_redisplay();
+	rl_replace_line("", 0);
 	ft_putchar_fd('\n', STDOUT_FILENO);
-	//rl_replace_line("", MAXCOM);
 	rl_on_new_line();
 	rl_redisplay();
 }
